refactor: refactor from laminas-mvc-console to laminas-cli
This patch builds on #19 in order to leverage laminas-cli for the purpose of the CLI tooling, instead of laminas-mvc-console.

It does so by refactoring CompileController into CompileCommand, removing the `bin/phly-blog` binary introduced by an earlier patch, extracting configuration to a `ConfigProvider` and updating the `Module` class to consume it, and adding laminas-cli configuration for the new command.