fix: do not mark Compiler as EventsManagerAware, but rather EventsCapable
laminas-mvc creates an initializer for classes marked EventsManagerAware.
Since initializers run AFTER delegators, if you attempt to add any listeners via a delegator factory, they then get overwritten when a new EM instance is injected by the initializer.

Additionally, this patch modifies the `Compiler` to optionally accept an EM instance to the constructor, and the `CompilerFactory` to pull one from the container when creating an instance.